+++ Client/window_menu.cpp
@@ -25,7 +25,7 @@ Event MenuWindow::processChar(TCPClient& client, const KeyPack& key, const std::
 Event MenuWindow::goUp() {
     selectedOption--;
     if (selectedOption < 0) {
-        selectedOption = options.size() - 1;
+        selectedOption = (int)options.size() - 1;
     }
     selectOption();
     return Event{};
